+++ ProjectX/Secondary.c
@@ -4637,8 +4637,8 @@ void DispSecAmmoLevels( void )
 	int16	x, y;
 	int16	Ammo;
 
-	x = (d3dappi.szClient.cx / 2 ) + ( 16 * 4 );
-	y = d3dappi.szClient.cy - ( d3dappi.szClient.cy / 5 ) + 6;
+	x = ((int16)d3dappi.szClient.cx / 2 ) + ( 16 * 4 );
+	y = (int16)d3dappi.szClient.cy - ( (int16)d3dappi.szClient.cy / 5 ) + 6;
 
 	Ammo = SecondaryAmmo[ Ships[ WhoIAm ].Secondary ];
